GPU3D: start fresh implementation of geometry engine

+++ src/common/polygon.h
@@ -5,5 +5,5 @@
 
 struct Polygon {
     Vertex* vertex_list = nullptr;
-    u8 size = 0;
+    int size = 0;
 };
